packages/core/pyproject.toml
[project]
name = "reconly-core"
version = "1.0.0"
description = "Reconly Core - CLI and content processing"
requires-python = ">=3.10"
authors = [
    { name = "Reconly Team", email = "hello@reconly.eu" }
]
license = { text = "AGPL-3.0" }
readme = "README.md"

dependencies = [
    "anthropic>=0.18.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "youtube-transcript-api>=0.6.0",
    "feedparser>=6.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "lxml>=5.0.0",
    "jinja2>=3.1.0",
    "openai>=1.0.0",
    "structlog>=24.1.0",
    "pgvector>=0.3.0",  # PostgreSQL vector similarity search
    "psycopg2-binary>=2.9.0",  # PostgreSQL driver
    "tiktoken>=0.7.0",  # Token counting for text chunking
    "numpy>=1.24.0",  # Array operations for embeddings
]

[project.scripts]
reconly = "reconly_core.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reconly_core"]
